+++ changelog.md
@@ -10,6 +10,7 @@
 ## Fixes:
 * **[Factions]** AI would never buy artillery units for the frontline - fixed
 * **[Factions]** Removed the F-111 unit from the NATO desert storm faction. (Recruiting it would cause crashes in DCS, since it is not a valid unit) 
+* **[Units]** Fixed SPG_Stryker_M1128_MGS not being in db
 
 # 2.3.2